+++ src/containerapp/azext_containerapp/custom.py
@@ -489,7 +489,15 @@ def create_containerapp(cmd,
                         workload_profile_name=None,
                         termination_grace_period=None,
                         secret_volume_mount=None,
-                        environment_type="managed"):
+                        environment_type="managed",
+                        source=None,
+                        repo=None,
+                        token=None,
+                        branch=None,
+                        context_path=None,
+                        service_principal_client_id=None,
+                        service_principal_client_secret=None,
+                        service_principal_tenant_id=None):
     raw_parameters = locals()
 
     containerapp_create_decorator = ContainerAppPreviewCreateDecorator(
@@ -542,7 +550,8 @@ def update_containerapp_logic(cmd,
                               registry_server=None,
                               registry_user=None,
                               registry_pass=None,
-                              secret_volume_mount=None):
+                              secret_volume_mount=None,
+                              source=None):
     raw_parameters = locals()
 
     containerapp_update_decorator = ContainerAppPreviewUpdateDecorator(
@@ -588,7 +597,8 @@ def update_containerapp(cmd,
                         workload_profile_name=None,
                         termination_grace_period=None,
                         no_wait=False,
-                        secret_volume_mount=None):
+                        secret_volume_mount=None,
+                        source=None):
     _validate_subscription_registered(cmd, CONTAINER_APPS_RP)
 
     return update_containerapp_logic(cmd=cmd,
@@ -619,7 +629,8 @@ def update_containerapp(cmd,
                                      workload_profile_name=workload_profile_name,
                                      termination_grace_period=termination_grace_period,
                                      no_wait=no_wait,
-                                     secret_volume_mount=secret_volume_mount)
+                                     secret_volume_mount=secret_volume_mount,
+                                     source=source)
 
 
 def show_containerapp(cmd, name, resource_group_name, show_secrets=False):